Relocating `findFiles` to be able to test it.
